+++ mdsplus/javascope/DataCached.java
@@ -1,14 +1,14 @@
-/* $Id: DataCached.java,v 1.2 2002/04/26 13:14:29 jgk Exp $ */
+/* $Id: DataCached.java,v 1.3 2002/07/26 10:55:56 manduchi Exp $ */
 import java.io.Serializable;
 
 public class DataCached implements Serializable
  {
     String expression;
     String experiment;
-    int    shot;
+    long    shot;
     Object data;
     
-    DataCached(String expression, String experiment, int shot, Object data)
+    DataCached(String expression, String experiment, long shot, Object data)
     {
         this.expression = expression;
         this.shot = shot;
@@ -16,7 +16,7 @@ public class DataCached implements Serializable
         this.data = data;
     }
     
-    public boolean equals(String expression, String experiment, int shot)
+    public boolean equals(String expression, String experiment, long shot)
     {
         return ( this.expression.equals(expression) && 
                  this.shot == shot &&
